AccountCreator fixes GitHub signup #10517
